agregar libros con foto operativo

+++ Recuperacion/Practica_3B_libreria/Libreria/admin/vistas/vista_admin.php
@@ -5,23 +5,19 @@ if (isset($_POST["btnAgregar"])) {
     $error_referencia_agre = $_POST["referencia_agre"] == "" || !is_numeric($_POST["referencia_agre"]) || $_POST["referencia_agre"] < 0;
 
     if (!$error_referencia_agre) {
-        $respuesta=consumir_servicios_REST(DIR_SERV."/repetido_insert/libros/referencia/".$_POST["referencia_agre"],"GET");
-        $json=json_decode($respuesta,true);
-        if(!$json)
-        {
+        $respuesta = consumir_servicios_REST(DIR_SERV . "/repetido_insert/libros/referencia/" . $_POST["referencia_agre"], "GET");
+        $json = json_decode($respuesta, true);
+        if (!$json) {
             session_destroy();
-            die(error_page("Práctica Rec 3B","<h1>Práctica Rec 3B</h1><p>Sin respuesta oportuna de la API</p>"));  
+            die(error_page("Práctica Rec 3B", "<h1>Práctica Rec 3B</h1><p>Sin respuesta oportuna de la API</p>"));
         }
 
-        if(isset($json["error_bd"]))
-        {
+        if (isset($json["error_bd"])) {
             session_destroy();
-            consumir_servicios_REST(DIR_SERV."/salir","POST",$datos_env);
-            die(error_page("Práctica Rec 3B","<h1>Práctica Rec 3B</h1><p>".$json["error_bd"]."</p>"));
+            consumir_servicios_REST(DIR_SERV . "/salir", "POST", $datos_env);
+            die(error_page("Práctica Rec 3B", "<h1>Práctica Rec 3B</h1><p>" . $json["error_bd"] . "</p>"));
         }
-        $error_usuario=$json["repetido"];
-
-      
+        $error_usuario = $json["repetido"];
     }
 
 
@@ -33,29 +29,27 @@ if (isset($_POST["btnAgregar"])) {
     $error_portada_agre = $_FILES["portada_agre"]["name"] != "" && ($_FILES["portada_agre"]["error"] || !$array_nombre || !getimagesize($_FILES["portada_agre"]["tmp_name"]) || $_FILES["portada_agre"]["size"] > 750 * 1024);
     $error_form_agre = $error_referencia_agre || $error_titulo_agre || $error_autor_agre || $error_descripcion_agre || $error_precio_agre || $error_portada_agre;
     if (!$error_form_agre) {
-            
-        $datos_env_insert["referencia_agre"]=$_POST["referencia_agre"];
-        $datos_env_insert["titulo_agre"]=$_POST["titulo_agre"];
-        $datos_env_insert["autor_agre"]=$_POST["autor_agre"];
-        $datos_env_insert["descripcion_agre"]=$_POST["descripcion_agre"];
-        $datos_env_insert["precio_agre"]=$_POST["precio_agre"];
 
-        $respuesta=consumir_servicios_REST(DIR_SERV."/insertar_libro","POST",$datos_env_insert);
-        $json=json_decode($respuesta,true);
-        if(!$json)
-        {
+        $datos_env_insert["referencia_agre"] = $_POST["referencia_agre"];
+        $datos_env_insert["titulo_agre"] = $_POST["titulo_agre"];
+        $datos_env_insert["autor_agre"] = $_POST["autor_agre"];
+        $datos_env_insert["descripcion_agre"] = $_POST["descripcion_agre"];
+        $datos_env_insert["precio_agre"] = $_POST["precio_agre"];
+
+        $respuesta = consumir_servicios_REST(DIR_SERV . "/insertar_libro", "POST", $datos_env_insert);
+        $json = json_decode($respuesta, true);
+        if (!$json) {
             session_destroy();
-            die(error_page("Práctica Rec 3B","<h1>Práctica Rec 3B</h1><p>Sin respuesta oportuna de la API</p>"));  
-        }
-    
-        if(isset($json["error_bd"]))
-        {
-            session_destroy();
-            consumir_servicios_REST(DIR_SERV."/salir","POST",$datos_env);
-            die(error_page("Práctica Rec 3B","<h1>Práctica Rec 3</h1><p>".$json["error_bd"]."</p>"));
+            die(error_page("Práctica Rec 3B", "<h1>Práctica Rec 3B</h1><p>Sin respuesta oportuna de la API</p>"));
         }
 
-            
+        if (isset($json["error_bd"])) {
+            session_destroy();
+            consumir_servicios_REST(DIR_SERV . "/salir", "POST", $datos_env);
+            die(error_page("Práctica Rec 3B", "<h1>Práctica Rec 3</h1><p>" . $json["error_bd"] . "</p>"));
+        }
+
+
 
         $_SESSION["accion"] = "Libro agregado con éxito";
         // realizo la inserccion de la foto
@@ -69,17 +63,20 @@ if (isset($_POST["btnAgregar"])) {
             @$var = move_uploaded_file($_FILES["portada_agre"]["tmp_name"], "../images/" . $nombre_nuevo);
 
             if ($var) {
-                try {
 
-                    $consulta = "update libros set portada=? where referencia=?";
-                    $sentencia = $conexion->prepare($consulta);
-                    $sentencia->execute([$nombre_nuevo,  $ultm_refe]);
-                    $sentencia = null;
-                } catch (PDOException $e) {
+                $datos_env_act["portada"] = $nombre_nuevo;
+                $datos_env_act["referencia"] = $ultm_refe;
+                $respuesta = consumir_servicios_REST(DIR_SERV . "/actualizar_foto", "PUT", $datos_env_act);
+                $json = json_decode($respuesta, true);
+                if (!$json) {
+                    session_destroy();
+                    die(error_page("Práctica Rec 3", "<h1>Práctica Rec 3</h1><p>Sin respuesta oportuna de la API</p>"));
+                }
+
+                if (isset($json["error_bd"])) {
                     if (file_exists("../images/" . $nombre_nuevo))
                         unlink("../images/" . $nombre_nuevo);
-                    $sentencia = null;
-                    $conexion = null;
+
                     $mensaje = "Usuario insertado con éxito pero con la imagen por defecto por un problema en la BD del servidor";
                 }
             } else {
